added javadocs fixed javadocs comments on p1

+++ p1_datastruktur/testing/collections/SortingTest.java
@@ -11,14 +11,20 @@ import org.junit.Test;
 import java.util.Arrays;
 
 /**
- * Created by Sebastian Börebäck on 2016-02-08.
+ * Sorting test class. Test the diffrent sorting methods.
+ * @author Sebastian Börebäck on 2016-02-08.
  */
 public class SortingTest {
 
 
-	private ArrayList<Double> list;
+	private ArrayList<Integer> list;
 	private double[] dbArr;
 
+	/**
+	 * Setup the test with a ArrayList of Integers and
+	 * a double array. both given random values
+	 * @throws Exception
+	 */
 	@Before
 	public void setUp() throws Exception {
 		//setup the list
@@ -29,12 +35,16 @@ public class SortingTest {
 		int[] temp = Utility.randomArray(size, 1, 100);
 		for (int i = 0; i < temp.length; i++) {
 			dbArr[i] = temp[i];
-			list.add((double) temp[i]);
+			list.add(temp[i]);
 		}
 
 	}
 
 
+	/**
+	 * Test Quicksort algorithm on the double array
+	 * @throws Exception
+	 */
 	@Test
 	public void testQuickSort() throws Exception {
 		//Im testing
@@ -50,6 +60,10 @@ public class SortingTest {
 
 	}
 
+	/**
+	 * Test Mergesort algorithm on a Integer Arraylist
+	 * @throws Exception
+	 */
 	@Test
 	public void testMergeSort() throws Exception {
